+++ src/test/kotlin/blackjack/domain/card/CardTest.kt
@@ -1,8 +1,5 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.card.Card
-import blackjack.domain.card.Denomination
-import blackjack.domain.card.Shape
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
