+++ src/test/kotlin/blackjack/domain/card/DenominationTest.kt
@@ -1,6 +1,5 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.card.Denomination
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
